+++ benchmark/src/jvmMain/kotlin/io/github/fukusaka/keel/benchmark/ServerHttpNioBenchmark.kt
@@ -30,6 +30,7 @@ object ServerHttpNioBenchmark : EngineBenchmark {
             config = IoEngineConfig(
                 threads = threads,
                 loggerFactory = benchmarkLoggerFactory(),
+                allocator = benchmarkAllocator(config),
             ),
         )
         val (connectorConfigure, tlsCloseable) = serverHttpConnectorConfig(config)