+++ benchmark/src/linuxMain/kotlin/io/github/fukusaka/keel/benchmark/ServerHttpEpollBenchmark.kt
@@ -25,6 +25,7 @@ object ServerHttpEpollBenchmark : EngineBenchmark {
             config = IoEngineConfig(
                 threads = threads,
                 loggerFactory = benchmarkLoggerFactory(),
+                allocator = benchmarkAllocator(config),
             ),
         )
         val (connectorConfigure, tlsCloseable) = serverHttpConnectorConfig(config)