+++ benchmark/src/macosMain/kotlin/io/github/fukusaka/keel/benchmark/ServerHttpNwConnectionBenchmark.kt
@@ -22,6 +22,7 @@ object ServerHttpNwConnectionBenchmark : EngineBenchmark {
             config = IoEngineConfig(
                 threads = threads,
                 loggerFactory = benchmarkLoggerFactory(),
+                allocator = benchmarkAllocator(config),
             ),
         )
         val (connectorConfigure, tlsCloseable) = serverHttpConnectorConfig(config)